AgenticAI/simple_chatbot_mcp_langgraph/backend/requirements.txt
pydantic
langchain
langgraph
langchain-core
langchain-community
langchain-groq
python-dotenv
arxiv
wikipedia
fastapi
uvicorn
tavily-python
structlog
ujson
mcp
